Release 0.3.2

+++ argo-nagios-ams-publisher.spec
@@ -10,7 +10,7 @@
 %endif
 
 Name:           argo-nagios-ams-publisher
-Version:        0.3.1
+Version:        0.3.2
 Release:        1%{mydist}
 Summary:        Bridge from Nagios to the ARGO Messaging system
 
@@ -57,13 +57,13 @@ install --directory --mode 755 $RPM_BUILD_ROOT/%{_localstatedir}/run/%{name}/
 %files -f INSTALLED_FILES
 %defattr(-,root,root,-)
 %config(noreplace) %{_sysconfdir}/%{name}/ams-publisher.conf 
+%config(noreplace) %{_sysconfdir}/%{name}/metric_data.avsc
 %dir %{python_sitelib}/%{underscore %{name}}
 %{python_sitelib}/%{underscore %{name}}/*.py[co]
 %defattr(-,nagios,nagios,-)
 %dir %{_localstatedir}/log/%{name}/
 %attr(0755,nagios,nagios) %dir %{_localstatedir}/run/%{name}/
-%dir %{_localstatedir}/spool/%{name}/metrics/
-%dir %{_localstatedir}/spool/%{name}/alarms/
+%dir %{_localstatedir}/spool/%{name}/
 
 %post
 %if 0%{?el7:1}
@@ -83,8 +83,8 @@ if [ "$1" = 0 ]; then
 %if 0%{?el7:1}
 %systemd_preun ams-publisher.service
 %else
-  /sbin/service ams-publisher stop > /dev/null 2>&1
-  /sbin/chkconfig --del ams-publisher 
+	/sbin/service ams-publisher stop > /dev/null 2>&1
+	/sbin/chkconfig --del ams-publisher 
 %endif
 fi
 exit 0
@@ -94,21 +94,26 @@ exit 0
 %systemd_postun_with_restart ams-publisher.service
 %endif 
 if [ "$1" = 0 ]; then
-  rm -rf %{_localstatedir}/run/%{name}/
+	test -d  %{_localstatedir}/run/%{name}/ && rm -rf %{_localstatedir}/run/%{name}/
 fi
 exit 0
 
 %pre
 if ! /usr/bin/id nagios &>/dev/null; then
-    /usr/sbin/useradd -r -m -d /var/log/nagios -s /bin/sh -c "nagios" nagios || \
+	/usr/sbin/useradd -r -m -d /var/log/nagios -s /bin/sh -c "nagios" nagios || \
         logger -t nagios/rpm "Unexpected error adding user \"nagios\". Aborting installation."
 fi
 if ! /usr/bin/getent group nagiocmd &>/dev/null; then
-    /usr/sbin/groupadd nagiocmd &>/dev/null || \
+	/usr/sbin/groupadd nagiocmd &>/dev/null || \
         logger -t nagios/rpm "Unexpected error adding group \"nagiocmd\". Aborting installation."
 fi
 
 %changelog
+* Thu Nov  8 2018 Daniel Vrcic <dvrcic@srce.hr> - 0.3.2-1%{?dist}
+- ARGO-1429 Improved msg counter stats for probe testing purposes
+- ARGO-1408 Ensure correct permissions on pidfile directory
+- ARGO-1348 Descriptive error in case delivery cache tool is called with queue
+  path not specified in configs
 * Tue Jun 19 2018 Daniel Vrcic <dvrcic@srce.hr> - 0.3.1-1%{?dist}
 - ARGO-1250 Inspection local socket is left with root permissions 
 - ARGO-1147 AMS publisher to add optional field
